refactor(rules): colocate trim-history with enforce-tdd
trim-history has a single consumer (enforce-tdd) and a rule-specific
type (HistoryWindow). Moving it out of rules/utils/ removes the false
signal that it is a shared helper.
